pyproject.toml
[project]
name = "karan-bot"
version = "0.1.0"
description = "Telegram bot with LangGraph, OpenAI, Chroma, ElevenLabs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic-settings>=2.5.2",
  "python-dotenv>=1.0.1",
  "langchain-openai>=0.2.2",
  "langchain-chroma>=0.1.0",
  "chromadb>=0.5.4",
  "langgraph>=0.2.43",
  "pytest-cov>=7.0.0",
  "langgraph-checkpoint-sqlite>=1.0.0",
  "python-telegram-bot>=21.6",
  "elevenlabs>=1.9.0",
  "tenacity>=9.0.0",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pillow>=11.3.0",
  "opentelemetry-sdk==1.35.0",
  "opentelemetry-exporter-otlp>=1.27.0",
  "opentelemetry-instrumentation-logging>=0.48b0",
  "prometheus-client>=0.20.0",
  "redis>=5.0",
  "SQLAlchemy>=2.0",
  "psycopg[binary]==3.2.10",
  "alembic==1.16.4",
  "aiohttp-retry==2.9.1",
  "amqp==5.3.1",
  "antlr4-python3-runtime==4.9.3",
  "appdirs==1.4.4",
  "atomicwrites==1.4.0",
  "atpublic==6.0.1",
  "billiard==4.2.1",
  "celery==5.5.3",
  "click-didyoumean==0.3.1",
  "click-plugins==1.1.1.2",
  "click-repl==0.3.0",
  "configobj==5.0.9",
  "databricks-sdk==0.58.0",
  "dictdiffer==0.9.0",
  "diskcache==5.6.3",
  "docker==7.1.0",
  "dpath==2.2.0",
  "dulwich==0.23.2",
  "dvc==3.61.0",
  "dvc-data==3.16.10",
  "dvc-http==2.32.0",
  "dvc-objects==5.1.1",
  "dvc-render==1.0.2",
  "dvc-studio-client==0.21.0",
  "dvc-task==0.40.2",
  "entrypoints==0.4",
  "fastapi==0.116.1",
  "flatten-dict==0.4.2",
  "flufl-lock==8.2.0",
  "funcy==2.0",
  "google-auth==2.40.3",
  "grandalf==0.8",
  "graphene==3.4.3",
  "graphql-core==3.2.6",
  "graphql-relay==3.2.0",
  "gto==1.7.2",
  "gunicorn==23.0.0",
  "hydra-core==1.3.2",
  "iterative-telemetry==0.0.10",
  "jsonpointer==2.1",
  "kombu==5.5.4",
  "mako==1.3.10",
  "matplotlib==3.10.0",
  "mlflow==3.1.1",
  "mlflow-skinny==3.1.1",
  "omegaconf==2.3.0",
  "opentelemetry-api==1.35.0",
  "opentelemetry-semantic-conventions==0.56b0",
  "orjson==3.10.18",
  "psycopg-binary==3.2.10",
  "pyasn1-modules==0.2.8",
  "pydot==4.0.1",
  "pygit2==1.18.0",
  "pygtrie==2.5.0",
  "pyls-spyder==0.4.0",
  "pyqt5==5.15.10",
  "pyqtwebengine==5.15.6",
  "rsa==4.9.1",
  "scmrepo==3.3.11",
  "setuptools==72.1.0",
  "shortuuid==1.0.13",
  "shtab==1.7.2",
  "sqlparse==0.5.3",
  "sqltrie==0.11.2",
  "starlette==0.47.1",
  "uvicorn==0.35.0",
  "vine==5.1.0",
  "voluptuous==0.15.2",
  "wheel==0.45.1",
  "zc-lockfile==3.0.post1",
]

[project.scripts]
karan-bot = "app.main:main"
karan-bot-telegram = "app.entrypoints.telegram_bot:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["src"]
